PDMatsExtras

PDMatsExtras.PSDMat

PDMatsExtras.WoodburyPDMat

PDMats.invquad!

PDMats.quad!

PDMatsExtras.submat

PDMatsExtras.PSDMat — Type

Positive semi-definite matrix together with a CholeskyPivoted factorization object.

PDMatsExtras.WoodburyPDMat — Type

WoodburyPDMat(
    A::AbstractMatrix{T}, D::Diagonal{T}, S::Diagonal{T},
) where {T<:Real}

Lazily represents matrices of the form

W = A * D * A' + S

D and S must have only non-negative entries.

Using this matrix type is a good idea if size(A, 1) > size(A, 2) as the structure in the matrix can be exploited to accelerate operations involving W's inverse [1], such as invquad, and it's determinant [2], such as logdet.

You probably don't want to use this matrix type if size(A, 1) < size(A, 2).

[1] - https://en.wikipedia.org/wiki/Woodburymatrixidentity [2] - https://en.wikipedia.org/wiki/Matrixdeterminantlemma

PDMats.invquad! — Method

invquad!(r::AbstractArray, a::AbstractPDMat, x::StridedMatrix)

Overwrite r with the value of the quadratic form defined by inv(a) applied columnwise to x

PDMats.quad! — Method

quad!(r::AbstractArray, a::AbstractPDMat, x::StridedMatrix)

Overwrite r with the value of the quadratic form defined by a applied columnwise to x

PDMatsExtras.submat — Method

submat(A::AbstractPDMat, inds)

Select a square submatrix of an AbstractPDMat specified by inds. Returns a AbstractPDMat of the same type.